Finished comments specs integration
