+++ spec/features/users_spec.rb
@@ -1,22 +1,7 @@
 require 'rails_helper'
 
-def visit_page_and_fill_in_form
-  visit new_user_registration_path
-  within('form') do
-    fill_in 'Name', with: 'john'
-    fill_in 'Email', with: 'john@gmail.com'
-    fill_in 'Password', with: 'password'
-    fill_in 'Password confirmation', with: 'password'
-  end
-end
-
-def sign_up_sign_out
-  click_button 'Sign up'
-  expect(page).to have_content('Welcome! You have signed up successfully.')
-  click_link 'Sign out'
-end
-
 RSpec.feature 'User' do
+  include Integration
   before(:each) do
     visit_page_and_fill_in_form
   end
@@ -53,6 +38,7 @@ RSpec.feature 'User' do
 end
 
 RSpec.feature 'User' do
+  include Integration
   before(:each) do
     visit_page_and_fill_in_form
   end
@@ -83,6 +69,7 @@ RSpec.feature 'User' do
 end
 
 RSpec.feature 'User' do
+  include Integration
   before(:each) do
     visit_page_and_fill_in_form
   end
@@ -104,6 +91,7 @@ RSpec.feature 'User' do
 end
 
 RSpec.feature 'User' do
+  include Integration
   before(:each) do
     visit_page_and_fill_in_form
   end
